Synpase Interface + implementation
The threshold of a neuron can be implemented via an inhibiting synapse. This
makes it possible to train the threshold via the normal learning algorithm.

The Interface is implemented by ConnectingSynapse and InhibitingSynapse.

ConnectingSynapse will be used to connect Neurons with a weighted Synapse.
InhibitingSynapse will be used as the inhibiting synapse and does not link
to a Neuron.